+++ R/functions.R
@@ -1,8 +1,15 @@
-#' Maximum number of line requested
+#' Maximum number of lines requested in Analytics-API
 #'
 #' @return int
-MAX_LINES_PER_REQUEST <- function() {
+MAX_LINES_PER_REQUEST_ANALYTICS_API <- function() {
   100000
+}
+
+#' Maximum number of lines requested in Users-API
+#'
+#' @return int
+MAX_LINES_PER_REQUEST_USERS_API <- function() {
+  1000
 }
 
 #' get_test_credentials
@@ -14,12 +21,14 @@ MAX_LINES_PER_REQUEST <- function() {
 get_test_credentials <- function() {
   tryCatch(
     {
-      piwik_pro_credentials <-list(
+      piwik_pro_credentials <- list(
         client_id = Sys.getenv("R_PIWIK_client_id"),
-        client_secret =  Sys.getenv("R_PIWIK_client_secret"),
-        url =  Sys.getenv("R_PIWIK_url"),
-        default_site =  Sys.getenv("R_PIWIK_default_site"),
-        alternative_site =  Sys.getenv("R_PIWIK_alternative_site")
+        client_secret = Sys.getenv("R_PIWIK_client_secret"),
+        url = Sys.getenv("R_PIWIK_url"),
+        default_site = Sys.getenv("R_PIWIK_default_site"),
+        alternative_site = Sys.getenv("R_PIWIK_alternative_site"),
+        client_id_users = Sys.getenv("R_PIWIK_client_id_users"),
+        client_secret_users = Sys.getenv("R_PIWIK_client_secret_users")
       )
       return(piwik_pro_credentials)
     },
@@ -30,12 +39,10 @@ get_test_credentials <- function() {
       return(list())
     }
   )
-
-
 }
 
 skip_if_missing_credentials <- function(credentials) {
-  if(!exists('client_id', credentials)) {
+  if (!exists("client_id", credentials)) {
     testthat::skip("No credentials for testing - please read the documentation... ")
   }
 }